+++ Engine_Master_UPC/ModuleDescriptors.cpp
@@ -34,7 +34,7 @@ bool ModuleDescriptors::init()
 void ModuleDescriptors::preRender()
 {
     // For now only the SRV heap is having deferred releases since it's the only one used for textures
-	UINT lastCompletedFrame = app->getModuleD3D12()->getLastCompletedFrame();
+	UINT lastCompletedFrame = (UINT)app->getModuleD3D12()->getLastCompletedFrame();
 	for (int i = 0; i < m_defferedDescriptors.size(); ++i) {
 
         if (lastCompletedFrame > m_defferedDescriptors[i].frame)
